Create extractCertsFromP12
Here's an explanation of what the script does:

The script first checks if it has the required command line arguments (the name of the p12 file and the password to unlock it).
It defines the names of the output files for the key, client certificate and CA chain.
It uses the openssl pkcs12 -info command to check if the p12 file includes a key, client certificate and CA chain, and prints a message for each.
It uses the openssl pkcs12 command to extract the key, client certificate and CA chain to separate files, using the -nocerts, -nokeys and -cacerts options respectively. The -passin option is used to provide the password for unlocking the p12 file.
It prints the names of the output files for the key, client certificate and CA chain.
Note that this script assumes that the p12 file includes only one key, client certificate and CA chain. If there are multiple entries of each type in the p12 file, the script may need to be modified to extract the correct one(s).